refactor(ds): stop storing dsfolder + sync task to load schema
