+++ recoco/apps/demarches_simplifiees/admin.py
@@ -4,8 +4,8 @@ from django.db.models import JSONField, QuerySet
 from django.http import HttpRequest
 from django_json_widget.widgets import JSONEditorWidget
 
-from .models import DSFolder, DSMapping, DSResource
-from .tasks import load_ds_resource_schema, update_or_create_ds_folder
+from .models import DSMapping, DSResource
+from .services import load_ds_resource_schema
 
 
 @admin.register(DSResource)
@@ -26,62 +26,6 @@ class DSResourceAdmin(admin.ModelAdmin):
             self.message_user(
                 request,
                 f"Tâche déclenchée pour la resource '{ds_resource.name}'.",
-                messages.SUCCESS,
-            )
-
-
-@admin.register(DSFolder)
-class DSFolderAdmin(admin.ModelAdmin):
-    list_display = (
-        "dossier_id",
-        "ds_resource",
-        "recommendation",
-    )
-
-    list_filter = ("created",)
-
-    search_fields = (
-        "ds_resource",
-        "dossier_id",
-    )
-
-    readonly_fields = (
-        "content_hash",
-        "dossier_id",
-        "dossier_number",
-        "dossier_prefill_token",
-        "dossier_url",
-        "recommendation",
-        "state",
-    )
-
-    formfield_overrides = {
-        JSONField: {"widget": JSONEditorWidget},
-    }
-
-    actions = ("update_matching",)
-
-    ordering = ("-created",)
-    list_select_related = (
-        "ds_resource",
-        "recommendation__project",
-    )
-
-    @admin.action(description="Mettre à jour le matching dossier / démarche simplifiée")
-    def update_matching(self, request: HttpRequest, queryset: QuerySet[DSResource]):
-        for ds_folder in queryset:
-            if not ds_folder.recommendation_id:
-                self.message_user(
-                    request,
-                    f"Le dossier '{ds_folder.dossier_id}' n'a pas de recommandation liée.",
-                    messages.ERROR,
-                )
-                continue
-
-            update_or_create_ds_folder.delay(ds_folder.recommendation_id)
-            self.message_user(
-                request,
-                f"Tâche déclenchée pour le dossier '{ds_folder.dossier_id}'.",
                 messages.SUCCESS,
             )
