--- recoco/apps/demarches_simplifiees/serializers.py
@@ -1,12 +1,0 @@
-from rest_framework import serializers
-
-from .models import DSFolder
-
-
-class DSFolderSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = DSFolder
-        fields = [
-            "dossier_url",
-            "prefilled_count",
-        ]
