+++ recoco/apps/projects/templates/projects/project/fragments/tasks_inline/task_item.html
@@ -90,7 +90,7 @@
                 <p class="fr-text-mention--grey fr-text--sm fr-my-3v">
                     En cliquant ci-dessous, vous créerez un dossier en brouillon sur Demarches-Simplifiees.fr.
                     <br>
-                    <span x-text="task.ds_folder.prefilled_count > 1 ? `${task.ds_folder.prefilled_count} champs seront pré-remplis avec les informations déjà renseignées ici.` : '1 champ sera pré-rempli avec les informations déjà renseignées ici.'"></span>
+                    Des champs seront pré-remplis avec les informations déjà renseignées ici.
                 </p>
                 <a class="fr-btn fr-btn--secondary fr-btn--sm"
                    @click.stop=""
