+++ recoco/apps/projects/templates/projects/project/fragments/tasks_inline/task_item_embed.html
@@ -50,11 +50,7 @@
                 <p class="fr-text-mention--grey fr-text--sm fr-my-3v">
                     En cliquant ci-dessous, vous créerez un dossier en brouillon sur Demarches-Simplifiees.fr.
                     <br>
-                    {% if task.ds_folder.prefilled_count > 1 %}
-                        {{ task.ds_folder.prefilled_count }} champs seront pré-remplis avec les informations déjà renseignées ici.
-                    {% else %}
-                        1 champ sera pré-rempli avec les informations déjà renseignées ici.
-                    {% endif %}
+                    Certains champs seront pré-remplis avec les informations dont nous disposons sur le dossier.
                 </p>
                 <a class="fr-btn fr-btn--secondary fr-btn--sm"
                    @click.stop=""
